Move mtev.* functions to mtev.mtev
