+++ src/modules/Makefile.in
@@ -62,6 +62,7 @@ HEADERS=lua_mtev.h
 LUA_FILES=mtev/extras.lua mtev/timeval.lua mtev/HttpClient.lua \
 	mtev/hooks/http.lua \
 	mtev/lua_init.lua \
+	mtev/mtev.lua \
 	mtev/Api.lua \
 	mtev/Proc.lua
