+++ src/modules/lua-support/mtev/lua_init.lua
@@ -12,3 +12,5 @@ mtev.extras = require("mtev.extras")
 mtev.Api = require("mtev.Api")
 
 mtev.Proc = require("mtev.Proc")
+
+require("mtev.mtev")
